tests/wpt/web-platform-tests/css/css-backgrounds/parsing/background-repeat-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders: getComputedStyle().backgroundRepeat</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-repeat">
<meta name="assert" content="background-attachment repeat value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("background-repeat", "repeat-x");
test_computed_value("background-repeat", "repeat-y");
test_computed_value("background-repeat", "repeat");
test_computed_value("background-repeat", "space");
test_computed_value("background-repeat", "round");
test_computed_value("background-repeat", "no-repeat");

test_computed_value("background-repeat", "repeat space");
test_computed_value("background-repeat", "round no-repeat");
test_computed_value("background-repeat", "repeat repeat", "repeat");

// See background-computed.html for a test with multiple background images.
test_computed_value("background-repeat", "repeat-x, repeat-y, repeat", "repeat-x");
</script>
</body>
</html>
